requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
opencv-python==4.9.0.80
librosa==0.10.1
torch==2.2.2
torchvision==0.17.2
Pillow==10.2.0
soundfile==0.12.1